+++ lib/feature_pack/api/controller.rb
@@ -1,3 +1,0 @@
-class FeaturePack::API::Controller < ActionController::API
-
-end