+++ spec/fixtures/feature_packs/group_foo-001_foo/_group_metadata/controller.rb
@@ -1,3 +1,0 @@
-class FeaturePack::PrevineBrasilController <  FeaturePack::GroupController
-
-end