+++ spec/fixtures/feature_packs/group_foo-001_foo/feature_bar-001_bar/controller.rb
@@ -1,3 +1,0 @@
-class FeaturePack::PrevineBrasil::CapitacaoPonderadaController < FeaturePack::Controller
-
-end