+++ spec/fixtures/feature_packs/group_foo-001_foo/feature_bar-001_bar/routes.rb
@@ -1,2 +1,0 @@
-# resources :captacao_ponderada, only: [:index]
-get '/', to: 'capitacao_ponderada#index'